Cleanup - remove bogus off_t in C code
Buffers in memory are inherently limited to size_t.
off_t is for offsets in files, which can be larger than memory.
It's harder to see what the code is doing if it's confused.
Also fseek()/ftell() only use "long"; we shouldn't pretend otherwise.

Now I think we only silently drop bits on arithmetic overflow :).

Unfortunately I don't recall any compiler warnings even on 32-bit.
You'd hope that was because _FILE_OFFSET_BITS == 32.  But it's not,
because we get built with the same compiler flags as python.
It's as if the C compiler silently coerces from 64 to 32 bits.

On the bright side, _FILE_OFFSET_BITS == 64 is very much a thing that we
want.  As was pointed out to me on IRC.