hashindex: remove stray off_t
off_t is for file offsets, not in-memory sizes.
On 32-bit this avoids a whole lot of compiler complaints

warning: conversion to ‘size_t’ from ‘off_t’ may alter its value
[-Wconversion]

Bucket sizes are not large.  (There's a file-format limitation.
We actually have a #define for it, although it's the wrong value and we
never refer to it.  So I removed that as well while I was there).